# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SGMA APP2 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/my:
    get:
      operationId: getMyTasks
      tags: [tasks]
      summary: List tasks assigned to the current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The user's assigned tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/reports/attachments/{attachmentId}:
    get:
      operationId: getTaskReportAttachment
      tags: [tasks]
      summary: Download a report attachment (assigned member or staff)
      security:
        - BearerAuth: []
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The file (base64)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskReportAttachmentData"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task's details (assigned member or staff)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}/reports:
    post:
      operationId: createTaskReport
      tags: [tasks]
      summary: Add a report or admin note to a task
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskReportInput"
      responses:
        "201":
          description: Updated task with the new report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TaskStatus:
      type: string
      enum: [NEW, IN_PROGRESS, WAITING_REVIEW, COMPLETED, POSTPONED, CANCELLED]
    TaskPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
    TaskReportType:
      type: string
      enum: [MEMBER_REPORT, ADMIN_NOTE]
    TaskParticipantRole:
      type: string
      enum: [SUPERVISOR, ASSIGNEE, SUPPORTER]
    TaskAssignee:
      type: object
      required: [id, taskId, userId, participantRole, isActive]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        participantRole:
          $ref: "#/components/schemas/TaskParticipantRole"
        userFullName:
          type: ["string", "null"]
        userAccount:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        userRole:
          type: ["string", "null"]
        isActive:
          type: boolean
        assignedAt:
          type: ["string", "null"]
          format: date-time
    TaskReportAttachment:
      type: object
      required: [id, reportId, fileName, mimeType, fileSize]
      properties:
        id:
          type: integer
        reportId:
          type: integer
        fileName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        createdAt:
          type: ["string", "null"]
          format: date-time
    TaskReportAttachmentData:
      type: object
      required: [id, fileName, mimeType, fileData]
      properties:
        id:
          type: integer
        fileName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        fileData:
          type: string
    CreateTaskReportFileInput:
      type: object
      required: [fileName, mimeType, fileData]
      properties:
        fileName:
          type: string
          minLength: 1
        mimeType:
          type: string
          minLength: 1
        fileData:
          type: string
          minLength: 1
    TaskReport:
      type: object
      required: [id, taskId, authorId, reportText, reportType, attachments]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: ["string", "null"]
        reportText:
          type: string
        progressPercent:
          type: ["integer", "null"]
        reportType:
          $ref: "#/components/schemas/TaskReportType"
        createdAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/TaskReportAttachment"
    TaskItem:
      type: object
      required: [id, title, description, priority, status, createdById, assignees]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdById:
          type: integer
        createdByName:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
        latestReportAt:
          type: ["string", "null"]
          format: date-time
        latestReportExcerpt:
          type: ["string", "null"]
        latestProgress:
          type: ["integer", "null"]
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/TaskAssignee"
    TaskDetail:
      type: object
      required: [id, title, description, priority, status, createdById, assignees, reports]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdById:
          type: integer
        createdByName:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/TaskAssignee"
        reports:
          type: array
          items:
            $ref: "#/components/schemas/TaskReport"
    TasksResponse:
      type: array
      items:
        $ref: "#/components/schemas/TaskItem"
...
    CreateTaskReportInput:
      type: object
      required: [reportText]
      properties:
        reportText:
          type: string
          minLength: 1
        progressPercent:
          type: ["integer", "null"]
        reportType:
          $ref: "#/components/schemas/TaskReportType"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/CreateTaskReportFileInput"
...